third_party/blink/web_tests/platform/generic/http/tests/eventsource/workers/eventsource-bad-mime-type-expected.txt

[Worker] Test EventSource with an event-stream with incorrect mime-type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-bad-mime-type.js
PASS [Worker] es.readyState is es.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
